Improving the no-argument behavior of `emux`
Now when sessions exist and no argument is provided to `emux`, it will
attach to the most last active session and start the interactive session
chooser, rather than just printing the list of sessions.

This improves the flow to be more consistently interactive, while making
the choice more visual, and also has the benefit of not dirtying the
shell's history with specific session names. 🙌